Fix review findings and split main.py into modules
- Fix crash: return ([], {}) instead of [] on error paths in fetch_and_process_pull_requests
- Add per_page=100 to GitHub API pagination (3x fewer API calls)
- Add timeout=30 to all requests.get() calls missing it
- Add JQL injection prevention via issue key validation
- Add JiraClient.close() and context manager support
- Refactor fetch_user_details to use github_api_request (gains rate limit retry)
- Remove unused exception classes (MissingRepoError, FileOperationError)
- Remove duplicate exception handling in __main__ block
- Delete stale setup.py and run_integration_tests.py
- Split main.py (1868 lines) into common.py, github_api.py, jira_processing.py, output.py
- Fix env var isolation in JIRA client tests
- Fix missing jira_user in integration test Args
- Add tests for extract_issue_keys_from_pr, JQL sanitization, session management, error return type
