+++ src/gh_pulls_summary/main.py
@@ -7,68 +7,46 @@ import os
 import re
 import subprocess
 import sys
-import time
 from typing import Any, cast
-from urllib.parse import quote
 
 import argcomplete
-import requests
 
-from gh_pulls_summary.jira_client import JiraClient, JiraClientError
-
-# Configuration
-GITHUB_API_BASE = "https://api.github.com"
-
-
-def get_github_headers(token: str | None = None) -> dict:
-    """
-    Get GitHub API headers, optionally with authentication.
-
-    Args:
-        token: GitHub personal access token (optional)
-
-    Returns:
-        Dictionary of headers for GitHub API requests
-    """
-    headers = {
-        "Accept": "application/vnd.github+json",
-        "X-GitHub-Api-Version": "2022-11-28",
-    }
-
-    if token:
-        headers["Authorization"] = f"Bearer {token}"
-
-    return headers
-
-
-# Custom Exception Classes
-class MissingRepoError(Exception):
-    """Raised when repository information cannot be determined."""
-
-
-class GitHubAPIError(Exception):
-    """Raised when GitHub API requests fail."""
-
-    def __init__(self, message, status_code=None, response_text=None):
-        self.status_code = status_code
-        self.response_text = response_text
-        super().__init__(message)
-
-
-class RateLimitError(GitHubAPIError):
-    """Raised when GitHub API rate limit is exceeded."""
-
-
-class NetworkError(Exception):
-    """Raised when network-related errors occur."""
-
-
-class ValidationError(Exception):
-    """Raised when input validation fails."""
-
-
-class FileOperationError(Exception):
-    """Raised when file operations fail."""
+from gh_pulls_summary.common import (  # noqa: F401
+    GitHubAPIError,
+    NetworkError,
+    RateLimitError,
+    ValidationError,
+    get_github_headers,
+)
+from gh_pulls_summary.github_api import (  # noqa: F401
+    fetch_file_content,
+    fetch_issue_events,
+    fetch_pr_diff,
+    fetch_pr_files,
+    fetch_pull_requests,
+    fetch_reviews,
+    fetch_single_pull_request,
+    fetch_user_details,
+    get_authenticated_user_info,
+    github_api_request,
+)
+from gh_pulls_summary.jira_client import JiraClient, JiraClientError  # noqa: F401
+from gh_pulls_summary.jira_processing import (  # noqa: F401
+    create_jira_client,
+    extract_issue_keys_from_pr,
+    extract_jira_from_file_contents,
+    extract_jira_issue_keys,
+    extract_primary_jira_from_file_contents,
+    extract_primary_jira_from_metadata,
+    get_rank_for_pr,
+)
+from gh_pulls_summary.output import (
+    create_markdown_table_header,
+    create_markdown_table_row,
+    generate_timestamp,
+    parse_column_titles,
+    validate_sort_column,
+)
 
 
 def get_repo_and_owner_from_git():
@@ -252,815 +230,6 @@ def configure_logging(debug):
     )
 
 
-def github_api_request(
-    endpoint, params=None, use_paging=True, max_retries=3, headers=None
-):
-    """
-    Makes a GitHub API request and optionally handles pagination.
-    Returns all results across all pages if pagination is enabled.
-    Returns None if no results are found or an error occurs.
-    Automatically handles rate limiting by waiting and retrying.
-
-    Args:
-        endpoint: GitHub API endpoint path
-        params: Query parameters for the request
-        use_paging: Whether to handle pagination automatically
-        max_retries: Maximum number of retries for rate limiting
-        headers: HTTP headers to use for the request
-    """
-    if params is None:
-        params = {}
-
-    if headers is None:
-        headers = get_github_headers()
-
-    all_results = []
-    page = 1
-    last_results = None  # Fail-safe to detect duplicate results
-
-    while True:
-        if use_paging:
-            params["page"] = page
-        url = f"{GITHUB_API_BASE}{endpoint}"
-        logging.debug(f"Making API request to {url} with params {params}")
-
-        # Retry loop for rate limiting
-        retry_count = 0
-        while retry_count <= max_retries:
-            try:
-                response = requests.get(url, headers=headers, params=params)
-            except requests.exceptions.ConnectionError as e:
-                raise NetworkError(
-                    f"Network connection failed. Please check your internet connection and try again. Details: {e}"
-                )
-            except requests.exceptions.Timeout as e:
-                raise NetworkError(
-                    f"Request timed out. The GitHub API may be slow. Please try again. Details: {e}"
-                )
-            except requests.exceptions.RequestException as e:
-                raise NetworkError(
-                    f"Network error occurred while contacting GitHub API. Details: {e}"
-                )
-
-            # Handle rate limiting with automatic retry
-            if (
-                response.status_code == 403
-                and "X-RateLimit-Remaining" in response.headers
-                and response.headers["X-RateLimit-Remaining"] == "0"
-            ):
-                if retry_count >= max_retries:
-                    reset_time = response.headers.get("X-RateLimit-Reset", "unknown")
-                    raise RateLimitError(
-                        f"GitHub API rate limit exceeded after {max_retries} retries. "
-                        f"Rate limit will reset at {reset_time}. "
-                        f"Consider using a GitHub token to increase your rate limit (5000 requests/hour vs 60 requests/hour). "
-                        f"Use --github-token or set GITHUB_TOKEN environment variable with a personal access token."
-                    )
-
-                # Parse reset time and calculate wait duration
-                reset_timestamp = int(response.headers.get("X-RateLimit-Reset", 0))
-                current_timestamp = int(time.time())
-                wait_seconds = max(
-                    reset_timestamp - current_timestamp + 1, 1
-                )  # Add 1 second buffer
-
-                # Convert timestamps to human-readable format for logging
-                from datetime import datetime
-
-                reset_datetime = datetime.fromtimestamp(reset_timestamp)
-                current_datetime = datetime.fromtimestamp(current_timestamp)
-
-                logging.warning(
-                    f"Rate limit exceeded. "
-                    f"Current time: {current_timestamp} ({current_datetime.isoformat()}), "
-                    f"Reset time: {reset_timestamp} ({reset_datetime.isoformat()}), "
-                    f"Wait duration: {wait_seconds} seconds ({wait_seconds / 60:.1f} minutes) "
-                    f"(retry {retry_count + 1}/{max_retries})..."
-                )
-                time.sleep(wait_seconds)
-                retry_count += 1
-                continue  # Retry the request
-
-            # If we got here, we didn't hit rate limit, so break out of retry loop
-            break
-
-        if response.status_code == 401:
-            raise GitHubAPIError(
-                "GitHub API authentication failed. Please check your --github-token or GITHUB_TOKEN if set. "
-                "You may need to generate a new personal access token from GitHub Settings.",
-                status_code=401,
-                response_text=response.text,
-            )
-
-        if response.status_code == 404:
-            raise GitHubAPIError(
-                f"GitHub API endpoint not found: {endpoint}. "
-                f"Please verify the repository owner and name are correct.",
-                status_code=404,
-                response_text=response.text,
-            )
-
-        if response.status_code != 200:
-            raise GitHubAPIError(
-                f"GitHub API request failed with status {response.status_code}. "
-                f"Endpoint: {endpoint}. "
-                f"Response: {response.text}",
-                status_code=response.status_code,
-                response_text=response.text,
-            )
-
-        try:
-            results = response.json()
-        except ValueError as e:
-            raise GitHubAPIError(
-                f"Invalid JSON response from GitHub API. The service may be experiencing issues. Details: {e}"
-            )
-
-        if results == last_results:  # pragma: no cover
-            logging.warning(
-                f"Duplicate results detected for pages {page - 1} and {page}. This may indicate a GitHub API issue."
-            )
-            break
-
-        # Handle cases where the response is a dictionary
-        if isinstance(results, dict):
-            logging.debug(f"Response is a dictionary: {results}")
-            return results
-
-        # Handle cases where the response is a list
-        if not results or not use_paging:
-            all_results.extend(results)
-            break
-
-        all_results.extend(results)
-        last_results = results
-
-        page += 1
-
-    logging.debug(f"Fetched {len(all_results)} items from {endpoint}")
-    return all_results
-
-
-def fetch_pull_requests(owner, repo, github_token=None, review_requested_for=None):
-    """
-    Fetches all open pull requests for the specified repository.
-    If review_requested_for is specified, fetches all PRs and filters using Search API intersection.
-    This ensures consistent full PR objects regardless of filtering.
-    """
-    headers = get_github_headers(github_token)
-
-    # Always fetch all open PRs for consistent data structure
-    endpoint = f"/repos/{owner}/{repo}/pulls"
-    params = {"state": "open"}
-    logging.debug(f"Fetching pull requests for {owner}/{repo}")
-    all_prs = github_api_request(endpoint, params, headers=headers)
-
-    if not all_prs:
-        return []
-
-    # If review_requested_for specified, use Search API to filter
-    if review_requested_for:
-        # Use Search API to get PR numbers where review is requested
-        search_endpoint = "/search/issues"
-        query = (
-            f"is:pr is:open repo:{owner}/{repo} review-requested:{review_requested_for}"
-        )
-        search_params = {"q": query, "per_page": 100}
-        logging.debug(
-            f"Filtering pull requests for {owner}/{repo} with review-requested:{review_requested_for}"
-        )
-
-        # Collect all matching PR numbers from Search API
-        matching_pr_numbers = set()
-        page = 1
-        while True:
-            search_params["page"] = page
-            search_results = github_api_request(
-                search_endpoint, search_params, use_paging=False, headers=headers
-            )
-            if not search_results or "items" not in search_results:
-                break
-
-            items = search_results.get("items", [])
-            if not items:
-                break
-
-            for item in items:
-                matching_pr_numbers.add(item["number"])
-            page += 1
-
-        # Return intersection: PRs that are in both /pulls and search results
-        filtered_prs = [pr for pr in all_prs if pr["number"] in matching_pr_numbers]
-        logging.debug(
-            f"Filtered {len(all_prs)} PRs to {len(filtered_prs)} matching review-requested:{review_requested_for}"
-        )
-        return filtered_prs
-
-    return all_prs
-
-
-def fetch_single_pull_request(owner, repo, pr_number, github_token=None):
-    """
-    Fetches a single pull request by its number.
-    """
-    endpoint = f"/repos/{owner}/{repo}/pulls/{pr_number}"
-    headers = get_github_headers(github_token)
-    logging.debug(f"Fetching single pull request #{pr_number} for {owner}/{repo}")
-    return github_api_request(endpoint, use_paging=False, headers=headers)
-
-
-def fetch_issue_events(owner, repo, pr_number, github_token=None):
-    """
-    Fetches all issue events for a specific pull request.
-    """
-    endpoint = f"/repos/{owner}/{repo}/issues/{pr_number}/events"
-    headers = get_github_headers(github_token)
-    logging.debug(f"Fetching issue events for PR #{pr_number}")
-    return github_api_request(endpoint, headers=headers)
-
-
-def fetch_reviews(owner, repo, pr_number, github_token=None):
-    """
-    Fetches all reviews for a specific pull request.
-    """
-    endpoint = f"/repos/{owner}/{repo}/pulls/{pr_number}/reviews"
-    headers = get_github_headers(github_token)
-    logging.debug(f"Fetching reviews for PR #{pr_number}")
-    return github_api_request(endpoint, headers=headers)
-
-
-def fetch_user_details(username, github_token=None):
-    """
-    Fetches details for a specific GitHub user.
-    Returns None if user is not found (404 error).
-    """
-    url = f"{GITHUB_API_BASE}/users/{username}"
-    headers = get_github_headers(github_token)
-    logging.debug(f"Fetching user details for {username}")
-
-    try:
-        response = requests.get(url, headers=headers)
-    except requests.exceptions.ConnectionError:
-        raise NetworkError(
-            f"Network connection failed while fetching user details for {username}. Please check your internet connection and try again."
-        )
-    except requests.exceptions.Timeout:
-        raise NetworkError(
-            f"Request timed out while fetching user details for {username}. The GitHub API may be slow. Please try again."
-        )
-    except requests.exceptions.RequestException as e:
-        raise NetworkError(
-            f"Network error occurred while fetching user details for {username}. Details: {e}"
-        )
-
-    if response.status_code == 404:
-        logging.debug(f"User {username} not found (404), returning None")
-        return None
-    if (
-        response.status_code == 403
-        and "X-RateLimit-Remaining" in response.headers
-        and response.headers["X-RateLimit-Remaining"] == "0"
-    ):
-        reset_time = response.headers.get("X-RateLimit-Reset", "unknown")
-        raise RateLimitError(
-            f"GitHub API rate limit exceeded while fetching user {username}. "
-            f"Rate limit will reset at {reset_time}. "
-            f"Consider using a GitHub token to increase your rate limit."
-        )
-    if response.status_code != 200:
-        raise GitHubAPIError(
-            f"Failed to fetch user details for {username}. "
-            f"GitHub API returned status {response.status_code}. "
-            f"Response: {response.text}",
-            status_code=response.status_code,
-            response_text=response.text,
-        )
-
-    try:
-        return response.json()
-    except ValueError as e:
-        raise GitHubAPIError(
-            f"Invalid JSON response when fetching user details for {username}. The GitHub API may be experiencing issues. Details: {e}"
-        )
-
-
-def fetch_pr_files(owner, repo, pr_number, github_token=None):
-    """
-    Fetches the list of files changed in a specific pull request.
-    """
-    endpoint = f"/repos/{owner}/{repo}/pulls/{pr_number}/files"
-    headers = get_github_headers(github_token)
-    logging.debug(f"Fetching files for PR #{pr_number}")
-    files = github_api_request(endpoint, use_paging=True, headers=headers)
-    logging.debug(f"Files fetched for PR #{pr_number}: {files}")
-    return files
-
-
-def fetch_file_content(owner, repo, file_path, ref, github_token=None):
-    """
-    Fetches the raw content of a file from GitHub at a specific ref (branch/commit).
-
-    Args:
-        owner: Repository owner
-        repo: Repository name
-        file_path: Path to the file in the repository
-        ref: Branch name or commit SHA
-        github_token: GitHub authentication token
-
-    Returns:
-        File content as string, or None if fetch fails
-    """
-    # URL-encode the file path to handle special characters like ?
-    # Keep forward slashes unencoded as they are path separators
-    encoded_file_path = quote(file_path, safe="/")
-    endpoint = f"/repos/{owner}/{repo}/contents/{encoded_file_path}"
-    headers = get_github_headers(github_token)
-    headers["Accept"] = "application/vnd.github.raw"
-    params = {"ref": ref}
-
-    logging.debug(f"Fetching content for {file_path} at ref {ref}")
-
-    try:
-        url = f"{GITHUB_API_BASE}{endpoint}"
-        response = requests.get(url, headers=headers, params=params, timeout=30)
-
-        if response.status_code == 200:
-            return response.text
-        if response.status_code == 404:
-            logging.warning(f"File not found: {file_path} at ref {ref}")
-            return None
-        if response.status_code == 403:
-            logging.warning(f"Access denied or rate limit when fetching {file_path}")
-            return None
-        logging.warning(f"Failed to fetch {file_path}: status {response.status_code}")
-        return None
-
-    except Exception as e:
-        logging.warning(f"Error fetching file content for {file_path}: {e}")
-        return None
-
-
-def create_jira_client(args) -> JiraClient | None:
-    """
-    Create a JIRA client if JIRA configuration is provided.
-
-    Args:
-        args: Command line arguments
-
-    Returns:
-        JiraClient instance or None if JIRA is not configured
-
-    Raises:
-        ValueError: If --include-rank is specified but JIRA configuration is invalid
-        JiraClientError: If --include-rank is specified but JIRA connection fails
-    """
-    if not args.include_rank:
-        return None
-
-    # If rank is requested, JIRA must be properly configured
-    try:
-        client = JiraClient(
-            base_url=args.jira_url,
-            user=args.jira_user,
-            token=args.jira_token,
-            rank_field_id=args.jira_rank_field,
-        )
-        # Test the connection
-        client.test_connection()
-        logging.info("JIRA client initialized successfully")
-        return client
-    except ValueError as e:
-        raise ValueError(
-            f"JIRA configuration error: {e}\n"
-            f"Rank column requested (--include-rank) but JIRA is not properly configured."
-        )
-    except JiraClientError as e:
-        raise JiraClientError(
-            f"JIRA connection failed: {e}\n"
-            f"Rank column requested (--include-rank) but cannot connect to JIRA."
-        )
-
-
-def extract_jira_issue_keys(url_dict: dict, pattern: str) -> list[str]:
-    """
-    Extract JIRA issue keys from URLs dictionary.
-
-    Args:
-        url_dict: Dictionary of URL text to URL
-        pattern: Regex pattern to match issue keys
-
-    Returns:
-        List of unique issue keys found
-    """
-    if not url_dict:
-        return []
-
-    issue_keys = set()
-    try:
-        regex = re.compile(pattern)
-        for url in url_dict.values():
-            matches = regex.findall(url)
-            issue_keys.update(matches)
-    except re.error as e:
-        logging.warning(f"Invalid JIRA issue pattern: {e}")
-        return []
-
-    return sorted(issue_keys)
-
-
-def extract_primary_jira_from_metadata(
-    content: str,
-    patterns: list[str],
-    row_pattern: str,
-    search_depth: int,
-) -> list[str]:
-    """
-    Extract primary JIRA issues from metadata table.
-
-    Looks for a markdown table with a row matching the specified pattern, e.g.:
-    | **Feature / Initiative** | [PROJECT-1586](url) |
-
-    Args:
-        content: Text content to search (PR body or file content)
-        patterns: List of regex patterns to extract issue keys (e.g., [r"(PROJECT-\\d+)"])
-        row_pattern: Regex pattern to identify the metadata row (case-insensitive)
-        search_depth: Number of lines to search from the top (-1 for unlimited)
-
-    Returns:
-        List of JIRA issue keys found in metadata table, or empty list if none found
-    """
-    if not content or not patterns:
-        return []
-
-    # Limit lines based on search depth
-    if search_depth < 0:
-        lines = content.split("\n")
-    else:
-        lines = content.split("\n")[:search_depth]
-
-    # Compile the row pattern (case-insensitive)
-    try:
-        metadata_row_regex = re.compile(row_pattern, re.IGNORECASE)
-    except re.error as e:
-        logging.warning(f"Invalid metadata row pattern '{row_pattern}': {e}")
-        return []
-
-    for line in lines:
-        # Check if this line contains the metadata row marker
-        if metadata_row_regex.search(line):
-            # Collect all unique matches from all patterns
-            all_matches = []
-            for pattern in patterns:
-                try:
-                    regex = re.compile(pattern)
-                    matches = regex.findall(line)
-                    if matches:
-                        all_matches.extend([str(m) for m in matches])
-                except re.error as e:
-                    logging.warning(f"Invalid JIRA issue pattern '{pattern}': {e}")
-                    continue
-
-            if all_matches:
-                # Remove duplicates while preserving order
-                unique_issues = list(dict.fromkeys(all_matches))
-                logging.info(
-                    f"Found primary JIRA issues in metadata table: {', '.join(unique_issues)}"
-                )
-                return unique_issues
-
-    logging.debug("No primary JIRA issues found in metadata table")
-    return []
-
-
-def extract_primary_jira_from_file_contents(
-    file_contents: list[str],
-    patterns: list[str],
-    row_pattern: str,
-    search_depth: int,
-) -> list[str]:
-    """
-    Extract primary JIRA issues from metadata tables in file contents.
-
-    Searches each file for a markdown table with a row matching the specified pattern.
-    Returns the first match found across all files.
-
-    Args:
-        file_contents: List of file content strings to search
-        patterns: List of regex patterns to extract issue keys (e.g., [r"(PROJECT-\\d+)"])
-        row_pattern: Regex pattern to identify the metadata row (case-insensitive)
-        search_depth: Number of lines to search from the top of each file (-1 for unlimited)
-
-    Returns:
-        List of JIRA issue keys found in metadata table, or empty list if none found
-    """
-    if not file_contents or not patterns:
-        return []
-
-    for content in file_contents:
-        if not content:
-            continue
-
-        # Check this file's content for metadata table
-        primary_issues = extract_primary_jira_from_metadata(
-            content, patterns, row_pattern, search_depth
-        )
-        if primary_issues:
-            logging.info(
-                f"Found primary JIRA issues in file metadata table: {', '.join(primary_issues)}"
-            )
-            return primary_issues
-
-    logging.debug("No primary JIRA issues found in file metadata tables")
-    return []
-
-
-def extract_jira_from_file_contents(
-    file_contents: list[str], patterns: list[str]
-) -> list[str]:
-    """
-    Extract JIRA issue keys from full file contents using multiple patterns.
-
-    Searches the complete content of all files (not limited to first 50 lines).
-    Each pattern should contain a single capture group that extracts the issue identifier.
-
-    Args:
-        file_contents: List of file content strings to search (full file contents)
-        patterns: List of regex patterns to extract issue keys (e.g., [r"(PROJECT-\\d+)", r"(OTHERJIRA-\\d+)"])
-
-    Returns:
-        Sorted list of unique JIRA issue keys found across all patterns
-    """
-    if not file_contents or not patterns:
-        return []
-
-    issue_keys: set[str] = set()
-
-    for pattern in patterns:
-        try:
-            regex = re.compile(pattern)
-            for content in file_contents:
-                if content:
-                    matches = regex.findall(content)
-                    # Add all matches to the set (duplicates automatically filtered)
-                    issue_keys.update(str(match) for match in matches)
-        except re.error as e:
-            logging.warning(f"Invalid JIRA issue pattern '{pattern}': {e}")
-            continue
-
-    return sorted(issue_keys)
-
-
-def extract_issue_keys_from_pr(
-    file_contents: list[str],
-    issue_patterns: list[str],
-    pr_body: str | None,
-    metadata_row_pattern: str,
-    metadata_search_depth: int,
-) -> list[str]:
-    """
-    Extract JIRA issue keys from a PR without fetching metadata.
-
-    Extraction strategy (priority order):
-    1. First checks PR body metadata table (searching N lines for metadata row pattern)
-    2. Then checks file contents for metadata tables (searching N lines of each file)
-    3. Falls back to searching full file contents if no metadata tables found
-
-    Args:
-        file_contents: List of file content strings to search for JIRA issues
-        issue_patterns: List of regex patterns to extract issue keys
-        pr_body: PR body/description text (optional)
-        metadata_row_pattern: Regex pattern to identify metadata row (case-insensitive)
-        metadata_search_depth: Number of lines to search from top (-1 for unlimited)
-
-    Returns:
-        List of JIRA issue keys found
-    """
-    if not issue_patterns:
-        return []
-
-    # Priority 1: Try to extract primary issues from PR body metadata table
-    issue_keys = []
-    if pr_body:
-        primary_issues = extract_primary_jira_from_metadata(
-            pr_body, issue_patterns, metadata_row_pattern, metadata_search_depth
-        )
-        if primary_issues:
-            issue_keys = primary_issues
-            logging.debug(
-                f"Found JIRA issues in PR body metadata: {', '.join(primary_issues)}"
-            )
-            return issue_keys
-
-    # Priority 2: Try to extract primary issues from file contents metadata tables
-    if file_contents:
-        primary_issues = extract_primary_jira_from_file_contents(
-            file_contents, issue_patterns, metadata_row_pattern, metadata_search_depth
-        )
-        if primary_issues:
-            issue_keys = primary_issues
-            logging.debug(
-                f"Found JIRA issues in file metadata: {', '.join(primary_issues)}"
-            )
-            return issue_keys
-
-    # Priority 3: Fall back to extracting from full file contents
-    issue_keys = extract_jira_from_file_contents(file_contents, issue_patterns)
-    if issue_keys:
-        logging.debug(f"Found JIRA issues in file contents (no metadata): {issue_keys}")
-
-    return issue_keys
-
-
-def get_rank_for_pr(
-    jira_client: JiraClient | None,
-    issue_keys: list[str],
-    jira_metadata_cache: dict[str, dict[str, Any]],
-) -> tuple[str | None, set[str]]:
-    """
-    Get the highest priority rank for a PR based on its JIRA issues.
-
-    Uses pre-fetched JIRA metadata to avoid redundant API calls.
-
-    Filtering rules:
-    - Only include Feature and Initiative issue types
-    - If referenced issue is not Feature/Initiative, traverse hierarchy to find ancestor
-    - Prefer open issues (New, Backlog, In Progress, Refinement)
-    - Fall back to closed issues if no open issues with ranks are found
-    - For multiple issues, select the highest priority (lowest lexicographic rank)
-    - Replace pipe characters with underscores for markdown safety
-
-    Args:
-        jira_client: JIRA client instance
-        issue_keys: List of JIRA issue keys for this PR
-        jira_metadata_cache: Pre-fetched metadata for all issues (must include parent fields)
-
-    Returns:
-        Tuple of (rank_string, closed_issue_keys):
-        - rank_string: Rank value with issue key appended, or None
-        - closed_issue_keys: Set of JIRA keys that are in Closed status
-    """
-    if not jira_client or not issue_keys:
-        return None, set()
-
-    # Get metadata for this PR's issues from cache
-    metadata = {
-        key: jira_metadata_cache[key]
-        for key in issue_keys
-        if key in jira_metadata_cache
-    }
-
-    if not metadata:
-        return None, set()
-
-    # Status definitions
-    open_statuses = {"New", "Backlog", "In Progress", "Refinement"}
-    closed_statuses = {"Closed"}
-
-    # Track closed issues and separate rank tuples by status
-    closed_issue_keys = set()
-    open_rank_tuples = []
-    closed_rank_tuples = []
-
-    for issue_key, issue_data in metadata.items():
-        issue_type = jira_client.get_issue_type(issue_data)
-        issue_status = jira_client.get_issue_status(issue_data)
-        rank_value = jira_client.extract_rank_value(issue_data)
-
-        logging.debug(
-            f"{issue_key}: type={issue_type}, status={issue_status}, rank={rank_value}"
-        )
-
-        # Track if this issue is closed
-        if issue_status in closed_statuses:
-            closed_issue_keys.add(issue_key)
-
-        # Check if this is a Feature or Initiative
-        if issue_type in ["Feature", "Initiative"]:
-            if rank_value:
-                if issue_status in open_statuses:
-                    open_rank_tuples.append((rank_value, issue_key))
-                elif issue_status in closed_statuses:
-                    closed_rank_tuples.append((rank_value, issue_key))
-        else:
-            # Not a Feature/Initiative - traverse hierarchy using cache to find ancestor
-            logging.info(
-                f"{issue_key} is type '{issue_type}' (not Feature/Initiative), traversing hierarchy to find ancestor"
-            )
-            try:
-                ancestors = jira_client.get_ancestors(
-                    issue_key, metadata_cache=jira_metadata_cache
-                )
-                for ancestor in ancestors:
-                    ancestor_key = ancestor.get("key")
-                    if not ancestor_key:
-                        continue
-
-                    ancestor_type = jira_client.get_issue_type(ancestor)
-                    ancestor_status = jira_client.get_issue_status(ancestor)
-                    ancestor_rank = jira_client.extract_rank_value(ancestor)
-
-                    logging.debug(
-                        f"  Ancestor {ancestor_key}: type={ancestor_type}, status={ancestor_status}, rank={ancestor_rank}"
-                    )
-
-                    # Track if ancestor is closed
-                    if ancestor_status in closed_statuses:
-                        closed_issue_keys.add(ancestor_key)
-
-                    if ancestor_type in ["Feature", "Initiative"] and ancestor_rank:
-                        if ancestor_status in open_statuses:
-                            logging.info(
-                                f"  Found open {ancestor_type} ancestor {ancestor_key} with rank {ancestor_rank}"
-                            )
-                            open_rank_tuples.append((ancestor_rank, ancestor_key))
-                            break
-                        if ancestor_status in closed_statuses:
-                            logging.info(
-                                f"  Found closed {ancestor_type} ancestor {ancestor_key} with rank {ancestor_rank}"
-                            )
-                            closed_rank_tuples.append((ancestor_rank, ancestor_key))
-                            break
-            except Exception as e:
-                logging.warning(f"Failed to traverse hierarchy for {issue_key}: {e}")
-
-    # Prefer open issues, fall back to closed issues
-    valid_rank_tuples = open_rank_tuples if open_rank_tuples else closed_rank_tuples
-
-    if not valid_rank_tuples:
-        return None, closed_issue_keys
-
-    # Select highest priority (lowest lexicographic value)
-    # Empty ranks should be treated as lowest priority
-    def rank_sort_key(rank_tuple):
-        rank_value = rank_tuple[0]
-        if not rank_value or rank_value == "":
-            return "z" * 100  # Push empty to end
-        return rank_value
-
-    valid_rank_tuples.sort(key=rank_sort_key)
-    highest_priority_rank, issue_key = valid_rank_tuples[0]
-
-    # Replace pipe characters with underscores for markdown safety
-    # Append issue key for transparency
-    rank_string = f"{highest_priority_rank.replace('|', '_')} {issue_key}"
-
-    return rank_string, closed_issue_keys
-
-
-def fetch_pr_diff(owner, repo, pr_number, github_token=None):
-    """
-    Fetches the diff for a specific pull request using the GitHub API.
-    Returns the diff as a string.
-    """
-    if not pr_number:
-        return None
-
-    url = f"{GITHUB_API_BASE}/repos/{owner}/{repo}/pulls/{pr_number}"
-    headers = get_github_headers(github_token)
-    headers["Accept"] = "application/vnd.github.v3.diff"
-
-    try:
-        response = requests.get(url, headers=headers)
-    except requests.exceptions.ConnectionError:
-        raise NetworkError(
-            f"Network connection failed while fetching diff for PR #{pr_number}. Please check your internet connection and try again."
-        )
-    except requests.exceptions.Timeout:
-        raise NetworkError(
-            f"Request timed out while fetching diff for PR #{pr_number}. The GitHub API may be slow. Please try again."
-        )
-    except requests.exceptions.RequestException as e:
-        raise NetworkError(
-            f"Network error occurred while fetching diff for PR #{pr_number}. Details: {e}"
-        )
-
-    if response.status_code == 404:
-        raise GitHubAPIError(
-            f"Pull request #{pr_number} not found in repository {owner}/{repo}. "
-            f"Please verify the PR number and repository are correct.",
-            status_code=404,
-            response_text=response.text,
-        )
-    if response.status_code == 403:
-        raise GitHubAPIError(
-            f"Access denied when fetching diff for PR #{pr_number}. "
-            f"The repository may be private or you may have exceeded rate limits. "
-            f"Consider using a GitHub token with appropriate permissions.",
-            status_code=403,
-            response_text=response.text,
-        )
-    if response.status_code != 200:
-        raise GitHubAPIError(
-            f"Failed to fetch diff for PR #{pr_number}. "
-            f"GitHub API returned status {response.status_code}. "
-            f"Response: {response.text}",
-            status_code=response.status_code,
-            response_text=response.text,
-        )
-
-    return response.text
-
-
 def fetch_and_process_pull_requests(
     owner,
     repo,
@@ -1103,7 +272,7 @@ def fetch_and_process_pull_requests(
         pr = fetch_single_pull_request(owner, repo, pr_number, github_token)
         if pr is None:
             logging.error(f"Failed to fetch PR #{pr_number}")
-            return []
+            return [], {}
         prs = [pr]  # Wrap in a list for consistent processing
     else:
         # Fetch all PRs (filtered by review_requested_for if specified)
@@ -1111,7 +280,7 @@ def fetch_and_process_pull_requests(
         prs = fetch_pull_requests(owner, repo, github_token, review_requested_for)
         if prs is None:
             logging.error("Failed to fetch pull requests")
-            return []
+            return [], {}
 
     url_regex_compiled = None
     if url_from_pr_content:
@@ -1401,142 +570,6 @@ def fetch_and_process_pull_requests(
     return pull_requests, jira_issues
 
 
-def parse_column_titles(args):
-    """
-    Parses custom column titles from command line arguments.
-    Returns a dictionary with the final column titles to use.
-    """
-    default_titles = {
-        "date": "Date",
-        "title": "Title",
-        "author": "Author",
-        "changes": "Change Requested",
-        "approvals": "Approvals",
-        "urls": "URLs",
-        "rank": "RANK",
-    }
-    custom_titles = {}
-    if hasattr(args, "column_title") and args.column_title:
-        for entry in args.column_title:
-            if "=" in entry:
-                col, val = entry.split("=", 1)
-                col = col.strip().lower()
-                if col in default_titles:
-                    custom_titles[col] = val.strip()
-                else:
-                    logging.warning(
-                        f"Invalid column name '{col}' in --column-title. Valid columns: {', '.join(default_titles.keys())}"
-                    )
-    return {**default_titles, **custom_titles}
-
-
-def validate_sort_column(sort_column):
-    """
-    Validates the sort column and returns it in lowercase.
-    Raises ValidationError if invalid.
-    """
-    allowed_columns = [
-        "date",
-        "title",
-        "author",
-        "changes",
-        "approvals",
-        "urls",
-        "rank",
-    ]
-    sort_column = sort_column.lower()
-    if sort_column not in allowed_columns:
-        raise ValidationError(
-            f"Invalid sort column: '{sort_column}'. "
-            f"Valid options are: {', '.join(allowed_columns)}. "
-            f"Use --sort-column to specify a valid column name."
-        )
-    return sort_column
-
-
-def create_markdown_table_header(titles, url_column, rank_column):
-    """
-    Creates the markdown table header and separator rows.
-    Returns a tuple of (header_row, separator_row).
-    """
-    header = f"| {titles['date']} | {titles['title']} | {titles['author']} | {titles['changes']} | {titles['approvals']} |"
-    separator = "| --- | --- | --- | --- | --- |"
-
-    if url_column:
-        header = header + f" {titles['urls']} |"
-        separator = separator + " --- |"
-
-    if rank_column:
-        header = header + f" {titles['rank']} |"
-        separator = separator + " --- |"
-
-    return header, separator
-
-
-def create_markdown_table_row(pr, url_column, rank_column, jira_issues=None):
-    """
-    Creates a single markdown table row for a pull request.
-
-    Args:
-        pr: Pull request data dictionary
-        url_column: Whether to include URLs column
-        rank_column: Whether to include rank column
-        jira_issues: Dictionary mapping JIRA keys to their data (for synthetic entries)
-    """
-    # Handle synthetic JIRA entries (no PR number)
-    if pr["number"] is None and "jira_key" in pr:
-        # Synthetic JIRA entry: lookup JIRA data
-        jira_key = pr["jira_key"]
-        if jira_issues and jira_key in jira_issues:
-            jira_data = jira_issues[jira_key]
-            title_link = f"[{jira_data['title']}]({jira_data['url']})"
-        else:
-            title_link = f"[{jira_key}]()"
-        author_link = ""
-        approvals_text = ""
-        changes_text = ""
-    else:
-        # Regular PR entry
-        title_link = f"{pr['title']} #[{pr['number']}]({pr['url']})"
-
-        # Handle author
-        if pr["author_name"]:
-            author_link = f"[{pr['author_name']}]({pr['author_url']})"
-        else:
-            author_link = ""
-
-        # Handle reviews/approvals
-        if pr["reviews"] > 0:
-            approvals_text = f"{pr['approvals']} of {pr['reviews']}"
-        else:
-            approvals_text = ""
-
-        # Handle changes (always show for regular PRs, even if 0)
-        changes_text = str(pr["changes"])
-
-    row = f"| {pr['date']} | {title_link} | {author_link} | {changes_text} | {approvals_text} |"
-
-    if url_column:
-        if pr.get("pr_body_urls_dict") and pr["pr_body_urls_dict"]:
-            closed_keys = pr.get("closed_issue_keys", set())
-            url_links = []
-            for text, url in pr["pr_body_urls_dict"].items():
-                # Apply strikethrough to closed JIRA issues
-                if text in closed_keys:
-                    url_links.append(f"[~~{text}~~]({url})")
-                else:
-                    url_links.append(f"[{text}]({url})")
-            row = row + f" {' '.join(url_links)} |"
-        else:
-            row = row + " |"
-
-    if rank_column:
-        rank_value = pr.get("rank", "")
-        row = row + f" {rank_value} |"
-
-    return row
-
-
 def generate_markdown_output(args):
     """
     Generates Markdown output for the given list of pull requests.
@@ -1696,42 +729,11 @@ def generate_markdown_output(args):
         row = create_markdown_table_row(pr, url_column, rank_column, jira_issues)
         output.append(row)
 
+    # Close JIRA client session if it was created
+    if jira_client:
+        jira_client.close()
+
     return "\n".join(output)
-
-
-def generate_timestamp(current_time=None, generator_name=None, generator_url=None):
-    """
-    Returns the current timestamp in Markdown syntax, including the generator's name and link if provided.
-    """
-    from datetime import datetime, timezone
-
-    if current_time is None:
-        current_time = datetime.now(timezone.utc)
-    timestamp = current_time.strftime("**Generated at %Y-%m-%d %H:%MZ**")
-    if generator_name and generator_url:
-        timestamp += f" by [{generator_name}]({generator_url})"
-    elif generator_name:
-        timestamp += f" by {generator_name}"
-    timestamp += "\n"
-    return timestamp
-
-
-def get_authenticated_user_info(github_token=None):
-    """
-    Fetch the authenticated user's info from the GitHub API.
-    Returns (name, html_url) or (None, None) if not available.
-    """
-    headers = get_github_headers(github_token)
-    try:
-        resp = requests.get("https://api.github.com/user", headers=headers, timeout=5)
-        if resp.status_code == 200:
-            data = resp.json()
-            name = data.get("name") or data.get("login")
-            html_url = data.get("html_url")
-            return name, html_url
-    except Exception:  # pragma: no cover
-        pass
-    return None, None
 
 
 def main():
@@ -1832,36 +834,3 @@ if __name__ == "__main__":  # pragma: no cover
     except KeyboardInterrupt:
         print("\nOperation cancelled by user.", file=sys.stderr)
         sys.exit(130)  # Standard exit code for SIGINT
-    except ValidationError as e:
-        print(f"ERROR: Input validation failed. {e}", file=sys.stderr)
-        sys.exit(1)
-    except JiraClientError as e:
-        print(f"ERROR: JIRA error. {e}", file=sys.stderr)
-        sys.exit(1)
-    except RateLimitError as e:
-        print(f"ERROR: GitHub API rate limit exceeded. {e}", file=sys.stderr)
-        sys.exit(1)
-    except GitHubAPIError as e:
-        print(f"ERROR: GitHub API error. {e}", file=sys.stderr)
-        sys.exit(1)
-    except NetworkError as e:
-        print(f"ERROR: Network error. {e}", file=sys.stderr)
-        sys.exit(1)
-    except FileOperationError as e:
-        print(f"ERROR: File operation failed. {e}", file=sys.stderr)
-        sys.exit(1)
-    except MissingRepoError as e:
-        print(f"ERROR: {e}", file=sys.stderr)
-        print(
-            "Either provide --owner and --repo arguments, or run the command",
-            file=sys.stderr,
-        )
-        print("from within a Git repository with a GitHub remote.", file=sys.stderr)
-        sys.exit(1)
-    except Exception as e:
-        print(f"ERROR: An unexpected error occurred. {e}", file=sys.stderr)
-        print("If this error persists, please report it as a bug.", file=sys.stderr)
-        import traceback
-
-        traceback.print_exc()
-        sys.exit(1)
